refactor: move duplicate union extraction and validation to _schema_output module
Move _extract_duplicate_union_types and validate_openapi_schema from
schema_transforms.py to _schema_output.py to satisfy the new 1000-line
file size pre-commit check.
